+++ tests/Assistants/VectorStoreTests.cs
@@ -187,7 +187,7 @@ public partial class VectorStoreTests
         _associationsToRemove.RemoveAt(0);
 
         // Errata: removals aren't immediately reflected when requesting the list
-        Thread.Sleep(1000);
+        Thread.Sleep(2000);
 
         int count = 0;
         foreach (VectorStoreFileAssociation association in client.GetFileAssociations(vectorStore).GetAllValues())
@@ -227,7 +227,7 @@ public partial class VectorStoreTests
         _associationsToRemove.RemoveAt(0);
 
         // Errata: removals aren't immediately reflected when requesting the list
-        Thread.Sleep(1000);
+        Thread.Sleep(2000);
 
         PageCollection<VectorStoreFileAssociation> pages = client.GetFileAssociations(vectorStore);
         IEnumerator<PageResult<VectorStoreFileAssociation>> pageEnumerator = ((IEnumerable<PageResult<VectorStoreFileAssociation>>)pages).GetEnumerator();